+++ vegefoods.angularui/src/app/Interceptor/auth-interceptor.service.ts
@@ -1,4 +1,4 @@
-import { AuthService } from './auth.service';
+import { AuthService } from '../Services/auth.service';
 import { Injectable } from '@angular/core';
 import {
   HttpInterceptor,
